nvme: move dir commands into a plugin
There are too many top-level commands cluttering the space. Move the
dir commands into a plugin, e.g.,

  nvme dir-receive -> nvme dir receive
