+++ meson_options.txt
@@ -230,7 +230,7 @@ option(
 option(
   'plugins',
   type : 'array',
-  choices: ['amzn', 'config', 'dapustor', 'dell', 'dera',
+  choices: ['amzn', 'config', 'dapustor', 'dell', 'dera', 'dir',
             'exclusion', 'fdp', 'feat', 'huawei', 'ibm', 'id', 'innogrit', 'inspur',
             'intel', 'keys', 'lm', 'log', 'mangoboost', 'memblaze', 'micron', 'nbft',
             'netapp', 'ns', 'nvidia', 'nvme-mi', 'ocp', 'registry', 'resv', 'sandisk', 'scaleflux',
